# perf(trie-router): optimize and remove unnecessary processes