@EdamAme-x 2024-11-09
# Report

Revert (4): The bundle size is smaller, but we got strange benchmark results that one index access is slower than multiple index accesses, so we reverted.
```llvm
runtime: deno 2.0.5 (x86_64-pc-windows-msvc)

benchmark              avg (min … max) p75   p99    (min … top 1%)
-------------------------------------- -------------------------------
Once                    193.31 ps/iter 195.31 ps ▇  █
                (170.90 ps … 70.87 ns) 341.80 ps █▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁
Multi                   136.00 ps/iter 146.48 ps     █
                 (97.66 ps … 22.39 ns) 219.73 ps ▁▁▁▁█▁▁▁▄▁▁▁▂▁▁▁▂▁▁▁▁
```

## Bundle Size

19.34kb => 19.28kb (1kb = 1024b)
```llvm
gitpod@honojs-hono-0oouxesc90f:/workspace/hono/sandbox$ ls compare -l
total 40
-rw-r--r-- 1 gitpod gitpod 19288 Nov  9 07:09 after.js
-rw-r--r-- 1 gitpod gitpod 19344 Nov  9 07:11 before.js
```

## Benchmark

### In Deno
There was not much change.
(+0.03%)
```llvm
|  Framework       | Runtime | Average | Ping       | Query      | Body       |
| ---------------- | ------- | ------- | ---------- | ---------- | ---------- |
| after | deno | 41,225.407 | 50,386.78 | 38,903.5 | 34,385.94 |
| before | deno | 41,079.073 | 50,923.38 | 39,116.56 | 33,197.28 |
```