@yusukebe 2024-11-10
@EdamAme-x 

Have you ever tried to benchmark frameworks on Bun, not only Deno?