@EdamAme-x 2024-11-11
I'll benchmark with several approaches.