@EdamAme-x 2024-11-11
I measured the performance of router itself using node, bun, and mitata, but did not see any significant performance improvement that we could see.