@yusukebe 2024-11-12
@EdamAme-x 

Thanks. Anyway, this is a good improvement. I'll merge this now.